Refactor: dto 네이밍 변경
