+++ src/main/java/swyp/team5/greening/post/dto/request/CreatePostRequestDto.java
@@ -4,6 +4,7 @@ package swyp.team5.greening.post.dto.request;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
+import swyp.team5.greening.post.dto.PostContentDto;
 
 public record CreatePostRequestDto(
 
@@ -14,7 +15,7 @@ public record CreatePostRequestDto(
         Long categoryId,
 
         @NotEmpty
-        List<ContentDto> content
+        List<PostContentDto> content
 
 ) {
