+++ src/main/java/swyp/team5/greening/post/dto/request/UpdatePostRequestDto.java
@@ -2,6 +2,7 @@ package swyp.team5.greening.post.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import java.util.List;
+import swyp.team5.greening.post.dto.PostContentDto;
 
 public record UpdatePostRequestDto(
 
@@ -9,7 +10,7 @@ public record UpdatePostRequestDto(
         String title,
 
         @NotEmpty
-        List<ContentDto> content
+        List<PostContentDto> content
 
 ) {
